Update README.md; add warnings.
... You're gonna brick so many people.
I, and other developers, would recommend just not releasing this, to be fair - some going as far as to say that it should be taken down for being malicious...
but I'm going to at least make a very clear and pointed warning to you, the developers, as well as the users, that "the usual 3DS homebrew users" will be bricking their consoles and whining about it. Hopefully some words in the README will alleviate that issue a bit.